# Sentri/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sentri Control Room API",
    "summary": "Governance preflight and telemetry for hosted AI agents",
    "description": "Sentri governance preflight for ChatGPT GPT Actions and Gemini API tools. Call sentriInteract before any proposed action and proceed only when status is completed and result.authorized is true. Immediately before each downstream tool call, consume its signed permit with verifySentriPermit and execute only the exact verified action. Then call reportSentriOutcome with every observed action result and its permit. Authenticate with the configured Sentri bearer token. Never retry or bypass blocked actions.",
    "version": "0.3.1"
  },
  "servers": [
    {
      "url": "https://YOUR_PUBLIC_SENTRI_HOST",
      "description": "Replace with the HTTPS URL that reaches the local Sentri service"
    }
  ],
  "paths": {
    "/interact": {
      "post": {
        "summary": "Interact",
        "operationId": "sentriInteract",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InteractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SentriBearer": []
          }
        ],
        "x-openai-isConsequential": true
      }
    },
    "/approvals/{thread_id}": {
      "post": {
        "summary": "Approve",
        "operationId": "reviewSentriExecution",
        "security": [
          {
            "SentriBearer": []
          }
        ],
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Thread Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InteractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-openai-isConsequential": true
      }
    },
    "/outcomes": {
      "post": {
        "summary": "Report Outcome",
        "description": "Attach observed tool results to a previously authorized workflow DAG.",
        "operationId": "reportSentriOutcome",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutcomeReport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutcomeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SentriBearer": []
          }
        ],
        "x-openai-isConsequential": false
      }
    },
    "/execute": {
      "post": {
        "summary": "Execute Gateway",
        "description": "Execute authorized model calls so Sentri observes provider usage directly.",
        "operationId": "executeSentriGateway",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewayExecutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayExecutionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SentriBearer": []
          }
        ],
        "x-openai-isConsequential": true
      }
    },
    "/permits/verify": {
      "post": {
        "summary": "Verify Permit",
        "operationId": "verifySentriPermit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermitVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermitVerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SentriBearer": []
          }
        ],
        "x-openai-isConsequential": true
      }
    },
    "/audit": {
      "get": {
        "summary": "Audit",
        "operationId": "getSentriAudit",
        "security": [
          {
            "SentriBearer": []
          }
        ],
        "parameters": [
          {
            "name": "execution_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by execution ID; omit to return the latest events.",
              "title": "Execution Id"
            },
            "description": "Filter by execution ID; omit to return the latest events."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 500,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "properties": {
          "approved": {
            "type": "boolean",
            "title": "Approved"
          },
          "reviewer": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Reviewer"
          },
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "approved",
          "reviewer",
          "reason"
        ],
        "title": "ApprovalRequest"
      },
      "AuditResponse": {
        "properties": {
          "execution_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Id"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/TelemetryEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "events"
        ],
        "title": "AuditResponse"
      },
...
      "GatewayAuthorization": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/PlannedAction"
          },
          "permit": {
            "type": "string",
            "maxLength": 8192,
            "minLength": 32,
            "title": "Permit"
          }
        },
        "type": "object",
        "required": [
          "action",
          "permit"
        ],
        "title": "GatewayAuthorization"
      },
      "GatewayExecutionRequest": {
        "properties": {
          "execution_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Execution Id"
          },
          "thread_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Thread Id"
          },
          "authorizations": {
            "items": {
              "$ref": "#/components/schemas/GatewayAuthorization"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Authorizations"
          },
          "upstream_tools": {
            "items": {
              "$ref": "#/components/schemas/UpstreamToolReference"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Upstream Tools",
            "description": "Optional caller-observed tools that produced context for this gateway call. Raw arguments and outputs are intentionally excluded."
          }
        },
        "type": "object",
        "required": [
          "execution_id",
          "thread_id",
          "authorizations"
        ],
        "title": "GatewayExecutionRequest"
      },
      "GatewayExecutionResponse": {
        "properties": {
          "execution_id": {
            "type": "string",
            "title": "Execution Id"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "succeeded",
              "failed",
              "cancelled",
              "mixed"
            ],
            "title": "Status"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/GatewayActionResult"
            },
            "type": "array",
            "title": "Results"
          },
          "outcome_event_id": {
            "type": "string",
            "title": "Outcome Event Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "execution_id",
          "thread_id",
          "status",
          "results",
          "outcome_event_id",
          "message"
        ],
        "title": "GatewayExecutionResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InteractRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Message",
            "description": "User request from the host chat"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "host": {
            "type": "string",
            "enum": [
              "chatgpt",
              "gemini",
              "local"
            ],
            "title": "Host",
            "default": "local"
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/PlannedAction"
            },
            "type": "array",
            "maxItems": 50,
            "title": "Actions"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "InteractRequest"
      },
      "InteractResponse": {
        "properties": {
          "execution_id": {
            "type": "string",
            "title": "Execution Id"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "approval_required",
              "blocked",
              "error"
            ],
            "title": "Status"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "alerts": {
            "items": {
              "$ref": "#/components/schemas/RiskAlert"
            },
            "type": "array",
            "title": "Alerts"
          },
          "cost": {
            "$ref": "#/components/schemas/CostMetrics"
          },
          "approval_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Url"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "execution_id",
          "thread_id",
          "status"
        ],
        "title": "InteractResponse"
      },
      "OutcomeReport": {
        "properties": {
          "execution_id": {
            "type": "string",
            "minLength": 1,
            "title": "Execution Id"
          },
          "thread_id": {
            "type": "string",
            "minLength": 1,
            "title": "Thread Id"
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/ActionOutcome"
            },
            "type": "array",
            "minItems": 1,
            "title": "Outcomes"
          },
          "final_response": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Final Response",
            "description": "Host response used to calculate a hash; raw content is not persisted."
          },
          "reported_by": {
            "type": "string",
            "enum": [
              "chatgpt",
              "gemini",
              "local",
              "external",
              "sentri_gateway"
            ],
            "title": "Reported By",
            "default": "external"
          },
          "upstream_tools": {
            "items": {
              "$ref": "#/components/schemas/UpstreamToolReference"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Upstream Tools"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "execution_id",
          "thread_id",
          "outcomes"
        ],
        "title": "OutcomeReport"
      },
      "OutcomeResponse": {
        "properties": {
          "execution_id": {
            "type": "string",
            "title": "Execution Id"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "succeeded",
              "failed",
              "cancelled",
              "mixed"
            ],
            "title": "Status"
          },
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "completed_dag": {
            "additionalProperties": true,
            "type": "object",
            "title": "Completed Dag"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "execution_id",
          "thread_id",
          "status",
          "event_id",
          "completed_dag",
          "message"
        ],
        "title": "OutcomeResponse"
      },
      "PermitVerificationRequest": {
        "properties": {
          "permit": {
            "type": "string",
            "maxLength": 8192,
            "minLength": 32,
            "title": "Permit"
          },
          "action": {
            "$ref": "#/components/schemas/PlannedAction"
          },
          "consume": {
            "type": "boolean",
            "title": "Consume",
            "description": "Consume the single-use permit before executing the action.",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "permit",
          "action"
        ],
        "title": "PermitVerificationRequest"
      },
      "PermitVerificationResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "consumed": {
            "type": "boolean",
            "title": "Consumed"
          },
          "claims": {
            "additionalProperties": true,
            "type": "object",
            "title": "Claims"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "valid",
          "consumed",
          "claims",
          "message"
        ],
        "title": "PermitVerificationResponse"
      },
      "PlannedAction": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Id"
          },
          "tool": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Tool",
            "description": "Exact downstream tool or API name"
          },
          "operation": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Operation",
            "description": "Operation the tool is expected to perform"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          },
          "rationale": {
            "type": "string",
            "maxLength": 2000,
            "title": "Rationale",
            "default": "Requested by the host conversation"
          },
          "mutates_state": {
            "type": "boolean",
            "title": "Mutates State",
            "default": false
          },
          "data_classification": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Data Classification",
            "description": "Data labels such as public, internal, sensitive, or pii"
          }
        },
        "type": "object",
        "required": [
          "tool",
          "operation"
        ],
        "title": "PlannedAction"
      },
      "RiskAlert": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "severity": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Severity"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "action_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Id"
          },
          "hard_limit": {
            "type": "boolean",
            "title": "Hard Limit",
            "default": false
          },
          "requires_human": {
            "type": "boolean",
            "title": "Requires Human",
            "default": false
          },
          "detected_at": {
            "type": "string",
            "format": "date-time",
            "title": "Detected At"
          }
        },
        "type": "object",
        "required": [
          "code",
          "severity",
          "message"
        ],
        "title": "RiskAlert"
      },
...
    },
    "securitySchemes": {
      "SentriBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}